Add test generation to smooth free-slip spherical cases
We only want to run two cases per level for the spherical cases, so
there's a new "permutate" option that doesn't generate all
permutations of the configuration parameters.

The Makefile has a special case for the spherical cases, which will
submit a sentinel job that will then wait for the result of all the
sub-cases.