Update expected results due to code changes
Viscoplastic result has been set to running the new version of the
code, after convergence improved. This matches the old version of the
code on the same Firedrake installation.

Compressible result has been updated due to the differing initial
condition compared to the old version of code. This has been verified
to match with the same Firedrake installation.

Cylindrical result has been updated after the removal of Ttheta from
the Stokes solve, and the reduction of the convergence tolerance. This
is within 3 iterations of convergence compared to the original code,
which may be attributed to minor roundoff issues.

For all cases, the new results were generated with Docker image ID
e9c634eb3031, using the following component git hashes:
- Firedrake: 0ec02b2d8d1afd07ad716c520e815039235ed255
- UFL: 3c62318c8319a2da985f4bf38a536ac06316be30
- PETSc: 9364cb008b2aaa08d45c84b965f81718cf7649cb