First go at SU advection for Gplates demo
Copied code from the paper for the gplates demo and tried on
a simple advection and advection diffusion example. At the moment
the term doesn't seem to be having an effect if I leave it as is.
If I reduce the diffusivity it looks like there ought to be a
diffusivity factor in the Pe term but because gadopt was nondimensionalised(?)
this has been left out... If I add what I think is the right
diffusivity scaling then the solution completely changes forming
a diffuse ring like a bulls eye pattern...