Add test ready (?) for github actions
scalar_advection_diffusion_DH27.py now outputs files when run
(which is set off by the Makefile - I think I might have misunderstood
how to use this as I am generating a file for each test with only
one make file...)

Test 1: for 3 different peclet numbers and with/wout SU
there is a check that you get the same L2 norm of q at steady state.

Test 2: for PE = 0.25 and with/wout SU check that the order of convergence
is 2. For higher peclet numbers we don't see this convergence rate (more
like 0.5 for Pe = 50) which presumably is because you aware from the
asymptotic limit as SU adds extra diffusion and is thus solving a different
problem... see earlier commits for more explanation.