Introduce VariableMassInvPC
Turns out MassInvPC is not actually written for variable (in time)
viscosity: it only assembles the preconditioner matrix once and is thus
based on the viscosity from the initial condition.

Implement drop-in replacement that does update (in nonlinear iterations
and between solves). At first a simple implementation based on
AuxillaryOperatorPC was tried, but it introduces a different options
structure with no ksp directly under the prefix (so you would
need pcksp and an additional options layer: prefix_ksp_ksp_type etc.).
This version copies a fair bit of MassInvPC code, but since we plan
further extensions for augmented lagrangian approaches, I feel this will
work better rather than proposing an upstream extension of firedrake's
MassInvPC to deal with variable viscosity.

Also adds SPDAssembledPC, which is a version of AssembledPC that sets
the SPD MatOption. To be used for inner velocity block with gamg.
And adds "assembled_pc_gamg_coarse_eq_limit": 1000. These are the 2 "new
options" that are recommended for GAMG, but don't seem to affect much in
testing.