Add spherical analytical with long test framework (#26)
* Add spherical analytical cases

* Add count subcommand to analytical script

This counts up the number of cases that are run for a particular
configuration, used for determining the number of jobs to depend on.

* Add test generation to smooth free-slip spherical cases

We only want to run two cases per level for the spherical cases, so
there's a new "permutate" option that doesn't generate all
permutations of the configuration parameters.

The Makefile has a special case for the spherical cases, which will
submit a sentinel job that will then wait for the result of all the
sub-cases.

* Generalise Makefile to work for all spherical cases

* Add skeleton workflow for longtests

Because it depends on on-disk structure on Gadi, this won't support
concurrent runs. We can't cancel existing runs because we don't have a
reasonable way to cancel in-flight jobs yet.

* Refactor analytical parameter set generation

* Run spherical analytical using a bash script

* Adjust analytical longtests to use Gadi prefix images

* Move Taylor test case definitions to separate file

Because the Taylor tests were being directly imported from their
pytest files, pytest could only be run if the gadopt and firedrake
modules are present. This is a bit restrictive, because we're only
running comparisons on the test outputs, which shouldn't require
these.

* Add separate mark for long tests

* Reduce tolerance requirement for spherical analytical tests

* Allow skipping of regression tests

[skip tests]

---------

Co-authored-by: Rhodri Davies <rhodri.davies@anu.edu.au>