Don't remove nullspace inside massinvpc solve. (#43)
The mass matrix we use as an approximation of the Schur complement does
not actually have a nullspace, so when we use cg to invert it in order
to apply it as a preconditioner, we should not be removing the nullspace
inside the cg iteration. This in fact led to frequent (silently failing)
solves - add ['fieldsplit_1']['Mp_ksp_converged_reason'] = None to see -
either failing with definite pc, or (worse) failing to converge
within 10000 iterations. The latter is of course a bit of time waste
(the pressure mass cg iteration is rel. fairly cheap, so not as bad as
you might think). The fact that this solve (silently) fails also didn't
seem to impact the overall convergence of the Schur complement (outer
iteration).

Note that this only leaves out the nullspace in the pressure mass solve
(fieldsplit_1_ksp_Mp); the nullspace is still being removed in the outer
Schur iteration (fieldsplit_1_ksp)

Also reduce excessive vtu output in 2d_cylindrical.