Move nu_bar calculation to energy solver
Now Calculates nubar in energy solver and passes to advection term
through fields.

This should make it easier for plotting (i.e. just output the fields
value).

Still not general enough for supg probably. I think that maybe you want to
have the calculation of the modified test function happening in the
energy solver and then passing this modified test function to add extra
terms in...