Add SU advection and scalar advection-diffusion demo (#21)
* First go at SU advection for Gplates demo

Copied code from the paper for the gplates demo and tried on
a simple advection and advection diffusion example. At the moment
the term doesn't seem to be having an effect if I leave it as is.
If I reduce the diffusivity it looks like there ought to be a
diffusivity factor in the Pe term but because gadopt was nondimensionalised(?)
this has been left out... If I add what I think is the right
diffusivity scaling then the solution completely changes forming
a diffuse ring like a bulls eye pattern...

* Fix SU formula to be function of kappa

Gplates demo was a special case with kappa =1. Now should be the same as Donea
and Huerta reference book.

Rotating advection example is stabilised by SU (and also if specified kappa =0).
Rotating advecion scalar diffusion SU has an obvious effect with
kappa = 1e-5 (less obvious for kappa = 1e-3). Without SU (kappa = 1e-5) has
obvious stripes when the resolution is 40 x 40 cells. Adding SU smooths out these
stripes. At a resolution of 160x 160 cells the stripes are much less noticeable
however SU is still quite a diffuse solution...

* Fix keyword argument to 'diffusivity' not kappa

Actually made very little difference to runs before because
the grid peclet number was so high! i.e. SU was adding a lot of
diffusion

* Change to SU on/off switch through equations not fields

Also added two more examples based on Chapter 2
Steady transport problems of Finite Element Methods for
Flow problems - Donea and Huerta, 2003.

For the pure diffusion case even with P1 getting some initial
over/undershoots near the slot and the top boundary. Not exactly
sure what is causing them though they disappear using P1DG+limiter
(from thwaites code example)... As an aside P1dg+limiter
for the spiral advection diffusion example 'looks' qualitatively the
best in that you can still see the original pattern more than
P1+SU but it doesnt have any over/undershoots.

Probably should do a comparison using the Donea and Huerta book
which has analytical solutions for the steady problems.

* Rewrite Donea Fig2.7 example with a function for error testing

errors compared with Donea and Huerta 2003 equation 2.23 for
the steady advection diffusion solution with constant source term,
constant velocity and homogenous dirichlet bcs on both sides

running with a grid peclet number of 50 and initial mesh resolution of 5 cells gives these results
SU_upwind =  [0.16548618, 0.11845977, 0.08582658, 0.06341912, 0.0456253,  0.02835158, 0.01551193, 0.00813498]  # i.e set Beta(Pe) = 1.
SU = [0.16343463, 0.11554859, 0.0816918,  0.05753857, 0.03732994, 0.0180603, 0.00614359, 0.00170367]
noSU = [0.57030194, 1.08824244, 0.22152179, 0.09494632, 0.04086379, 0.01557628,  0.00484643, 0.00130424]

running with a grid peclet number of 0.25 and initial mesh resolution of 5 cells gives these results
SU_fromPe0pt25 = [0.01091291, 0.00276109, 0.00069343, 0.00017465]
noSU_fromPe0pt25 = [0.00863846, 0.00217787, 0.00054636, 0.00013744]
SU_upwind_fromPe0pt25 = [0.04179666, 0.02128963, 0.01080482, 0.00545314] # i.e. set Beta(Pe) =1

at higher grid peclet numbers (greater than about 10) both upwinding
and su are pretty much the same (because the Be(Pe) ->1 = upwinding)
and have lower errors than galerkin with no su (due to oscillations).

The convergence for SU and upwinding seems to be about 0.5 at this stage.
I couldnt find a theoretical reason why but it kind of makes sense
that you are not in the asymptotic limit because both methods add extra diffusion
based on the grid cell size, so as you refine the mesh you are effectively
changing the problem you are solving, compared with the analytical solution
that has a fixed specified diffusivity.

As the grid peclet number goes below 1 the convergence of SU and
galerkin both approaches 2 which makes sense for a P1 discretisation.
The convergence of the upwinding method approaches 1 which also makes
sense because you are no longer using a central difference type approximation.

* plotting convergce script

* Fix lint

* Add test ready (?) for github actions

scalar_advection_diffusion_DH27.py now outputs files when run
(which is set off by the Makefile - I think I might have misunderstood
how to use this as I am generating a file for each test with only
one make file...)

Test 1: for 3 different peclet numbers and with/wout SU
there is a check that you get the same L2 norm of q at steady state.

Test 2: for PE = 0.25 and with/wout SU check that the order of convergence
is 2. For higher peclet numbers we don't see this convergence rate (more
like 0.5 for Pe = 50) which presumably is because you aware from the
asymptotic limit as SU adds extra diffusion and is thus solving a different
problem... see earlier commits for more explanation.

* Modify testing for SU

This fixes up the Makefile to run the different cases in parallel, and
control the set of parameters from a single place (the test file). The
expected results are coalesced into a single file to slightly reduce clutter.

* Fix case name in Makefile

* Fix path resolution in SU test

* Tidying code for scalar advection example

* More code tidying changes

* Tidying code for demo 2.7 from Donea and Huerta 2003

* Tidying code

* Move nu_bar calculation to energy solver

Now Calculates nubar in energy solver and passes to advection term
through fields.

This should make it easier for plotting (i.e. just output the fields
value).

Still not general enough for supg probably. I think that maybe you want to
have the calculation of the modified test function happening in the
energy solver and then passing this modified test function to add extra
terms in...

* fix lint

* Update demos to use G-ADOPT energy equation

* Change makefile for test summary

* Update comments and function space names

* Fixes for Angus' suggestions

---------

Co-authored-by: Angus Gibson <angus@agibson.me>
Co-authored-by: Rhodri Davies <rhodrid@Rhodris-Mac-Pro.lan>
Co-authored-by: Rhodri Davies <Rhodri.Davies@anu.edu.au>