Modify free surface formulation
This commit:
- Removes the additional fields syntax and adds optional arguments for
the free surface in the stokes solver. These get used to turn the
free surface stress boundary term on in stokes_integrators and momentum
directly so the user is not responsible for this.

- Nondimensionalises the equations so now in terms of dimensionless time
and the box dimensions (and free surface height) are scaled by the depth
of the mantle. This means the solver residuals are now about 10 orders of
magnitude lower than before when everything was dimensional.

- Updates the expected testing results that now have a much lower
absolute magnitude due to nondimensionalising the equations (though the
relative errors are still converging at the correct rate).