Fix analytical (delta) and spiegelman tests with initialise() hack...
In the last commit I introduced a new method for StokesSolver
which delayed initialising the residual terms to allow for the free
surface. Since spiegelamn and the analytical (delta) cases both modify
F in the runscript this causes an error because they have not been set
of. The quick fix is just to call initialise in the runscript but
is this really what we want?