Change stokes integrators to break equation setup into methods
setup bcs as a method so that free surface can overwrite.
also setup equations separately so that you can add the eta terms
and also provide a free surface id.

This removes all the extra initiliase() calls that were needed
in the previous commit to keep the spiegelman and analytical
comparisons working where F was modified in the runscript