Make inheritance in stokes integrators neater - Thanks Thomas!
After chatting through the code with Thomas we hopefully tidied it up
more. The main change was to follow Sia's advice and not have a
separate FreeSurfaceStokesEquations list. I think this makes sense
because really the equations are similar just there is a different
bc (which requires another equation...) We still have the separate
FreeSurfaceStokesSolver though because the solving method does change
quite a bit. The nice thing now is that both of the additional methods
setup_bcs() and setup_equations() are calling the same method from the
parent class so you are inheriting everything...

Also added another method to overload solve. This means that the original
StokesSolver doesn't need to know about the solution from the previous
timestep, like it currently is in the main branch.

Also some tidying up in syntax to try and keep it similar to the original
code.