Rework free surface again so that it is turned on through bcs
Following the meeting we agreed it would make more sense for the
free surface to `turn on' through specifying a keyword in the
boundary condition dictionary and not by calling a separate
FreeSurfaceStokesSolver.

I have removed the extra class and made small modifications to
StokesSolver. I think the current approach works well. We generate
a new dictionary for the free surface bc information if the 'free_surface'
keyword is found when looping over the original bcs. We then use the
new free_surface_dict to setup a free surface equation and variables
based on the boundary id.

This means we can deal with multiple free surfaces and I now have added
a working example to github actions based on the Kramer et al 2012 paper (Section 3.1.2).
Since the top and bottom free surfaces are identical the decay is almost
exactly the same (except for roundoff errors), i.e. top and bottom free surface
converge to analytical solution at second order. This included adding
an option for 'exterior_density' as well as dot(n,k) to make sure the orientation
is correct.