Fix velocity nullspace with a small absorption term
This adds a small absorption term bringing the vertical velocity
towards zero. This leads to correct convergence behaviour provided
that the term is not too small or too large.

If it is too small (i.e. alpha = 1e-3 * 2/dtf) you get the nullspace behaviour
where the average vertical velocity oscillates between positive and
negative values between timesteps and you lose convergence as the timestep
decreases.

If it is too large (for instance you dont scale the term as the timestep
gets smaller) then the absorption term dominates over the numerical
error and you lose convergence again.

For some reason pinning the vertical velocity by using a modified
strong dirichlet bc didnt work. It seemed to be applying in the
correct place (at the stationary point (lambda/4, D/2)) but then
the actual value was never zero and it led to a funny dipole pattern
blowing up from the location... Anyway, since pinning the bc
was also a hack and this other hack of a small absorption term seems
to work I think we can move on for now.