Try CI test for smooth cylindrical free surface
Also allowed nullspace to be valid for more than just u,p mixed
system.

Need the rotational nullspace to get velocity convergence (though
you can get the correct convergence if you do the rotation projection
correction at the end of the timesteps... )

Also the near nullspace translation option makes a big difference
to the FS0 number of iterations. Without the option you have about 150 and
with you have about 30, which is the same as without the free surface