Ice loading taylor test works (with forward working as well...)
main thing was not updating the ice load at every timestep by interpolation
because this probably is not adjointable and does not really make sense...
maybe you would need a control at every timestep if you wanted to get the
sensitivity to the ice load at that timestep...

Also making sure that the objective function is squared vertical
displacement.

Also slgihtly odd with constant for ramp. in the forward model
have to constant as is, but the adjoint needs domain=mesh. I thought
that this was not necessary/depreciated now... not sure...
maybe need to use Real function space but it didnt seem to help...
at the moment got hacky inheritance to define different iceloads
for inverse and forward. I think the solution is to have ice
load symbolically depend on an initial iceload function and then
vary through by a simple relationship like no change in time, or
linear change in time, or piecewise linear between multiple
unknown ice functions...

hopefully this also fixes the ci??? during the mesh resolution changes
I had changed dx -> self.dx but for the original 2d case these had
been 5km and 10km respectively. The ice loading disc tanh profile
used the 5km version so the displacement results were off by a bit.

I think this was the major change though there were small changes possibly with
making the output work in cylinder domains..