Add vertex only mesh surface outputs back in
Using the newer input ordering method to make sure that the vertex
only mesh output works ok in parallel.

Unfortunately curved surfaces have not been introduced for vertex only
mesh yet so I have gone back to using degree =1 for the cylinder (and
eventually spherical geometry on setonix). Not sure how important this is
yet...

You get identical answers for the greatest negative displacement
 down to round off (16 ish digits?) using the bc method and the vom
method for both the rectangle and cylinder geometry. Plus the minimum value
is under the ice load corresponding with the first entry in the nodes list.

For the cylinder when plotting displacement vs theta we will need
take the dot product with the normal vector to get the radial
vs tangential displacement. But I think it is more sensible to do this
later as postprocessing.