Remove VariableMassInvPC (#60)
* Remove VariableMassInvPC

The main reason the have our own version of MassInvPC was that MassInvPC
did not update for varying (between solves) viscosities. This
has been fixed in https://github.com/firedrakeproject/firedrake/pull/3372
and since then the options structure for the solvers beneath it has
changed: need to use pcksp with an additional ksp_ layer now. Thus
VariableMassInvPC is no longer a drop in-replacement for MassInvPC
as it was intended. Better to stay consistent with Firedrake therefore.