Convert GeodynamicalDiagnostics to accept a mixed function (#75)
Instead of splitting our mixed function `z` and then passing `u` and `p`
around (e.g. to `GeodynamicalDiagnostics`), it seems to make more sense
to provide it with `z` directly. We have precedent for this already in
`StokesSolver`, especially with the free surface fields coming in #51.

Additionally, instead of exposing some of the gory MPI internals in our
demos, I've pushed that into a `ux_max` diagnostic, which can be given a
surface ID. I think related to the discussion in #74, this would be
better off knowing what the "lateral" direction is, as it just makes the
cartesian assumption at the moment.