Add diagram showing relationship of demos (#84)
I've already put a `.pages` file in the demos directory that is used for
the table of contents that shows up on the documentation site. I think
that, and this diagram make more sense here because they can be updated
alongside the demos.

Also has links to the tutorials that show up in the hierarchy, plus
adjoint on the side (without a link until #80 goes in).

[skip tests]