Fix first implicit free surface case
Update to work with new iterative solver argument in main branch.

Also updated the expected results. I think some change in Firedrake
changed the expected error norm to the 4th significant figure.
I thought it initially it might be changing between Firedrake's mass inv preconditioner
and gadopt's but it didnt seem to be that. It only affects the iterative
cases so presumably it is not a change to the free surface code. I couldnt see
any obvious changes to my code before the CI broke...

I've checked that it qualitatively looks right when plotting free surface evolution
and crucially the second order convergence test still passes.