Make `cartesian` a mesh attribute (#99)
There are several places where we would like to know whether the domain
is Cartesian or not, and forgetting to specify this can cause silently
confusing behaviour. It seems to make sense that the mesh itself should
know whether it represents a Cartesian domain, and indeed a mesh is
present in all places where previously a `cartesian` kwarg would have
been required.

The usage of this approach is fairly simple:
```python
mesh = SomeMeshFunction(...)
mesh.cartesian = True # or False
```
From here on, the `vertical_component` or `upward_normal` functions can
query the mesh directly, and it doesn't have to be passed around as a
flag. In the case when the `cartesian` attribute is *not* present on the
mesh, I think this will fail with an `AttributeError`. Perhaps this is
better than silently making an assumption about the domain, but it could
certainly do with a more informative error message in this case.