Turn off constant jacobian with adapting dt!
Because dt is varying and is part of the LHS matrix need to make
sure that jacobian is updated! With this fixed the simulation reaches a
steady state with RaFS=10Ra (seems reasonable for mantle alpha=3e-5K^-1,
delta_T = 3000K) whereas before the free surface oscillated and
didnt reach a steady state.

The final steady state for temp/vel/pressure is the same as the base case
and  the free surface height is equivalent to a dynamic topography
calculation of the normal stress at the top surface for the base case
(rigid lid) as expected.