Jupytext demo for adjoint (#80)
This PR introduces the demo for adjoint. It also changes the forward
case in our adjoint box case to serve as a demo if possible. Note that
the inverse.py and taylor_test.py are as before, they only now work with
the new filename for the state variables from the forward run.

---------

Co-authored-by: Rhodri Davies <rhodri.davies@anu.edu.au>
Co-authored-by: Angus Gibson <angus@agibson.me>