Reduce no. of cores
CI longtest failed on a similar test case... previous fix was to reduce
no. of cores. N.b. because we are timesteppign to a steady state
the resolution is actually reduced by a factor of 4 compared with the
other scripts to try and make it faster... So maybe we don't need
such high number of cores