Add analytical viscoelastic loading test
Based on Cathles et al 2024 Equation 2. The test here is loading
instead of unloading so we don't have to run to isostatic equilibrium
to begin with.

Although zhong's method claims to be second order we only see first order
convergence. I believe this is due to errors in the free surface dominating
not errors in timestepping the constitutive equation, based on
experiments with velocity formulation (moresi).