Allow testing against several result sets
This is important for the 2d_cylindrical_TALA_DG case in particular,
at least for the moment. There's some kind of parallel reproducibility
bug where we get one of two reasonably similar results. I don't think
we really want to slacken the tolerances to the point where they're
both accepted for a single set of expected values, so this is the next
best thing.