Move stress definition and input viscosity to approximation (#121)
This moves the definition of stress to the approximation. It also
removes viscosity as an input argument to stokes solver instead being
passed through the approximation.

This should simplify the setup e.g. see Issue #111 and the PR #112 to
try and fix it.

I would like to implement something like this so that it is easier to
change the definition of stress for viscoelastic consitutive equations
which depend on time.

Closes #111 
Closes #112 

---------

Co-authored-by: Rhodri Davies <rhodri.davies@anu.edu.au>
Co-authored-by: Angus Gibson <angus@agibson.me>