Backend changes: Equations and terms (#113)
This set of changes makes `Equation` a dataclass that holds the
entire state for a given equation and implements terms as functions.

With the proposed changes, there is only one namespace per
equation (`Equation`), and terms accept an `Equation` instance as their
first argument to draw from that single namespace. Note that terms also
require a second argument: the trial function object. This choice
results from the fact that time steppers must evaluate trial functions
at different times, and so having the trial object inside the namespace
is troublesome.