Change ice load to expression (Thomas suggestion)
It changes the results a bit (<10%). Did some testing integrating ice load
over the surface with an expression vs a function and
because our quadrature degree is 6 on a coarse mesh get slightly more
accurate results using an expression as opposed to interpolating
the ice load into a P1 space. It's all about how well is the tanh
approximated by the grid/FEM discretisation/quadrature degree.
If you refine the grid then you get the same answer