Fix adjoint
Found that previous changes had broken adjoint. tracked it down
to the interpolate steps!

For some reason interpolate with split(solution) doesnt work. but
subfunctions does. maybe makes sense because this is an explicit rhs
term so probably just want to interpolate the values as opposed to
just expressing as an expression?