Viscoelastic with stress defined in approximation (#122)
This is another go at viscoelasticity based off #121. The motivation is
to tidy up the code so that the stress can be calculated through the
approximation. Hopefully this should be tidier than #117 and it makes
adding in the internal variable approach easier where the stress
definition is quite different.

It still passes the same tests as #117 (with the same values) : 1)
convergence tests using incremental displacement formulation for three
end members (elastic, viscoelastic and viscous) for an analytical
surface loading problem. 2) 

---------

Co-authored-by: Angus Gibson <angus@agibson.me>
Co-authored-by: Rhodri Davies <rhodri.davies@anu.edu.au>
Co-authored-by: william.scott1@anu.edu.au <wis15@LAPTOP-RL9CDOVP>
Co-authored-by: Dale Roberts <dale.roberts1@anu.edu.au>