Cleanup of unused tests and rationalisation of adjoint testing.  (#133)
On the adjoint side, we currently have:

demos/adjoint - a nice succinct demo, for recovery of an initial
condition in a square domain. There is a forward script and an
associated adjoint (including a Taylor test). Currently only Taylor test
is tested. My view is that we should test all aspects (including
minimisation up to perhaps 5 iterations) and I will update that today.

demos/cylindrical - a longer more challenging test, including a
non-linear rheology, in a cylinder - this is essentially the final
synthetic test from Sia's 2024 paper. My plan is to actually have the CI
run the Taylor test (from a checkpoint) and remove inverse.py - testing
it within the CI is simply not tractable (many many many hours on many
cores!) and since it is not tested, it ends up not compatible with
master either way.

tests/adjoint - this is quite similar to demos/adjoint, albeit with a
more complex functional. Taylor tests are calculated for all 4 terms in
the functional, which is currently part of the CI. We will keep this,
but again remove inverse.py, because it is untested and has already
become incompatible with master.

Note that both inverse scripts that I removed are archived on zenodo -
associated with Ghelichkhan et al. 2024 - so that we can recover the
versions used in that paper easily.

This will address https://github.com/g-adopt/g-adopt/issues/127 and most
of https://github.com/g-adopt/g-adopt/issues/126.

---------

Co-authored-by: Angus Gibson <angus@agibson.me>