Serialise startup of render jobs
jupytext calls nbconvert to use its ExecutePreprocessor which starts a
NotebookClient which starts a KernelManager through a
LocalProvisioner. Of course, KernelManager includes a
ConnectionFileMixin and tries to bind 5 random ports when it
provisions the manager, then writes this to a file for when the Client
actually starts. This is a big race condition, and means that the
automatic retry code in the Client for random ports never gets hit,
because the port is already specified in the connection file.

The solution herein isn't foolproof, but barring a big hang in the
initialisation of jupyter it should work. We serialise just the
startup of the render jobs through a lockfile which gives each process
1 second to start before releasing the lock.

[skip tests]