Run render jobs with namespaced networking
This is possibly a less hacky alternative to #147 for ensuring all the
jupyter kernels don't run into each other when they try to bind ports
on startup.

There are plusses and minuses to this approach. The environment needs
to be more carefully set up so that the namespace works, including
setting subuid and subgid. I think it's probably also the case that
the render jobs won't have external network access without setting up
a veth device in the namespace. Maybe that's not the worst thing -- we
get a fairly robust guarantee that we don't have any network
dependencies.