Generate optimisation checkpoints with collective ID (#151)
Mistaken change in 30a3388 (Use comm-free UUID for optimisation
checkpointing) broke parallel compatibility of optimisation
checkpointing. Of course, the checkpoints for the ROL vectors need to be
collective, so they all need to write to the same file with the same ID.
We can go back to Firedrake's _get_uid, and just use the communicator
associated with the function on the vector.