Backend changes: GenericTransportSolver (#114)
This second set of changes extracted from gh-102 proposes to reformat
the `energy_solver` module to expose both `AdvectionDiffusionSolver` and
`EnergySolver` classes. These changes do not impact the use of
`EnergySolver`. The motivation here is the fact that some of our demos
and tests set up scenarios that only require an advection or
advection-diffusion equation. Currently, we use
`BoussinesqApproximation` and `EnergySolver` in these scripts, but this
is not straightforward. In other words, we make it unnecessarily
complicated for a user to solve such problems using G-ADOPT. The
proposed changes solve this issue and allow additionally to solve a pure
diffusion problem.

Needs gh-113.

---------

Co-authored-by: Angus Gibson <angus@agibson.me>