Detect boundary ids based on info from the mesh (#149)
Use the underlying PETSc plex object (if it exists) to detect
Firedrake's boundary subdomain ids and return them in a namespace. The
actual `get_boundary_ids` function is in `utility.py`, the rest of the
changes are updating demos/tests to use it. Not ready for merge yet,
just want to see if it passes testing, I've only run the demos so far.

---------

Co-authored-by: Angus Gibson <angus@agibson.me>
Co-authored-by: Rhodri Davies <rhodri.davies@anu.edu.au>