Change su_diffusivity back to an on/off flag
PR #114 changed the api to allow users to specify an ```su_diffusivity```
This is maybe confusing because if a 'real' background diffusivity is given
then we should use this diffusivity used to calculate the grid peclet number for the
su advection scheme.

In the case where there is no diffusion term we just set kappa=0 since we
already have a hack to specify a (hopefully) small 1e-12 value for the
diffusivity anyway.