Use firedrake modules on Gadi for longtest (#172)
Works fine, just had to change that iteration component name to get it
to work. I was going to put the `module use`, `module load` steps in
`run_gadi.sh`, but I changed my mind as its a good idea to have that
separate to the repository so that we can test new firedrake branches as
well. Location and usage of `$GADOPT_SETUP` would be good to add to our
hypothetical internal docs page.

---------

Co-authored-by: Angus Gibson <angus@agibson.me>