Allow running demos independent of /usr/bin/time and tsp
Ideally we should be able to run
    make -C g-adopt/demos/mantle_convection/base_case check

And have it just work. When we are still running in the CI, it would
be nice to retain some progress information, particularly when jobs
start/finish.

Having taken a brief look at the tests, this could be more complex,
particularly with the structure of the analytical comparisons. I think
this could be changed to invoke make instead of tsp directly.