Allow running demos without task-spooler (#178)
Also gets rid of the requirement on `/usr/bin/time` which has different
behaviour on BSD (i.e. macOS).

Closes #175.
Closes #167.

# How it works
This adds a `rules.mk` in the base directory, which has methods for
running tests in `BATCH_MODE` (i.e. CI) or interactively, and uses
`mpiexec` or not, depending on the case. Both modes will print some
timing information. The important section is the `run-python` canned
recipe which depends on the following variables: `$ncpus`, `$category`,
`$PETSC_FLAGS`, `$exec_args` and expects the Python script to be the
first dependency (as `$<`). There's also a pattern rule which uses the
canned recipe. For example:

```makefile
params.log : category := mantle_convection
params.log: run-base_case
```

The `check` rules all use a relative path to find the test fixture, so
they will work with `make -C` from wherever.

---------

Co-authored-by: dsroberts <115434035+dsroberts@users.noreply.github.com>