Add unit test for solver parameter consistency
In response to a recent Firedrake regression where solver parameters
weren't being passed through to blocks on the tape. This is a bit
brute force, but we ensure that our artificial "gadopt_test" option is
processed by the options database when we play the tape forward by
evaluating the reduced functional.