Use correct type for optimisation parameters (#217)
The "Safeguard Size" and "Gradient Tolerance" parameters that are
present in the minimisation defaults are actually floats, but they are
written as ints. With a default build of ROL using boost::property_list,
this doesn't cause any issues. However, it's possible to build ROL using
Teuchos to manage the parameter lists, and it'll complain about type
issues.