Add github action to create memory profiling plots (#230)
This action is set to run on completion of the regression tests. It
grabs the last 10 `run-log` artefacts and runs this script:
https://github.com/g-adopt/memprof_plotter/blob/main/memprof_plotter/plotter.py
to create some plots and upload them as an artefact. Note that
`workflow_run` triggers only work if the workflow file is on the main
branch, so it won't do anything until the PR is merged. Here are my test
runs: https://github.com/dsroberts/actions_mess_around/actions.

Once its merged, the graphs won't be particularly useful until more
branches are created from after #221 was merged. The python script does
handle the case where the `memprof` table is not present in the
database.