Integrate `gadopt_hpc_helper` module with longtests (#260)
To address #174, I've come up with an optional add-on for G-ADOPT to
assist on deployment on HPC clusters (Currently at
https://github.com/dsroberts/gadopt_hpc_helper - will move to g-adopt).
There is no need for desktop G-ADOPT users to ever use it, however, I
plan to install it into our firedrake modules on Gadi and Setonix (and
any other system we end up running on one day). The idea is that it
detects what system you're on, and it has knowledge of what queues are
available and their properties, what scheduler is in use and how to run
parallel jobs. Two of the three entry points are demonstrated in this
change, the `gadopt_hpcrun` script, which sort of replaces `qsub` or
`mpiexec`, and the `gadopt_hpcrun_async` python function, which is the
preferred option for running many batch jobs out of a single python
script/make target like the analytical longtests. There is also a
`gadopt_hpcrun` function that is just wraps `gadopt_hpcrun_async` in its
own event loop so a user doesn't need to worry about dealing with
`asyncio`. Given a new firedrake patch release is close, I'll install
`gadopt_hpc_helper` into the next firedrake module, so the long tests
won't work (and therefore this PR shouldn't be merged) until that's
available on both Gadi and Setonix.

In the meantime, I'd appreciate any input on
https://github.com/dsroberts/gadopt_hpc_helper, as its going to be the
main driver behind the HPC testing now.